requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.8
decorator==5.1.1
distlib==0.3.6
dnspython==2.3.0
fastapi==0.96.0
ffmpy==0.3.0
filelock==3.12.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gradio==3.33.1
gradio_client==0.2.6
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.232
langchainplus-sdk==0.0.17
lazy_loader==0.2
librosa==0.10.0.post2
linkify-it-py==2.0.2
llvmlite==0.40.1
load-dotenv==0.1.0
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nulltype==2.3.1
numba==0.57.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pinecone-client==2.2.2
pipenv==2023.6.12
platformdirs==3.5.3
pooch==1.6.0
psycopg-binary==3.1.9
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyscopg2==66.0.2
python-dateutil==2.8.2
python-dotenv==0.15.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.1
semantic-version==2.10.0
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.5
SQLAlchemy==2.0.16
starlette==0.27.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
toolz==0.12.0
tqdm==4.65.0
twilio==8.4.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
firebase-admin==5.0.1
flask-migrate==3.1.0
plaid_python==14.0.0